:recycle: Fix Executor
