+++ Sources/Kyusu/Executor/Executor.swift
@@ -1,17 +1,11 @@
 //
 //  Executor.swift
-//  Commander
+//  Kyusu
 //
-//  Created by Yudai.Hirose on 2019/06/04.
+//  Created by Yudai Hirose on 2019/06/26.
 //
 
 import Foundation
-import SwiftShell
-
-public struct ExecutorInfo {
-    let path: Path
-    let command: Command
-}
 
 public protocol Executor {
     associatedtype Information
@@ -25,29 +19,6 @@ public enum ExecutorError: Error {
         switch self {
         case .unknown:
             return "Unexpected error from SwiftShell."
-        }
-    }
-}
-
-public struct TeapotCommandExecutor: Executor {
-    public init() {
-        
-    }
-    public func exec(information: ExecutorInfo) throws {
-        let command = information.command.joined(separator: " ") + " \(information.path)"
-        let output = main.run(bash: command)
-        switch output.succeeded {
-        case false:
-            switch output.error {
-            case .none:
-                print("[ERROR]🚫: " + ExecutorError.unknown.localizedDescription)
-                exit(1)
-            case .some(let error):
-                print("[ERROR]🚫: " + error.localizedDescription)
-                exit(2)
-            }
-        case true:
-            print(output.stdout)
         }
     }
 }
